OpenTitan fork of upstream Tock
This branch is the opentitan fork of upstream tock.  This branch will be
used for development, debug and testing of changes before upstreaming.

This branch currently has a number of changes that are not yet committed
to upstream.  They include:

alistair23's initial work to bring tock up to date with opentitan.
cfrantz's disabling of the current ePMP implementation.
cfrantz's refactoring of earlgrey_cw310 into a bin+lib crate.
cfrantz's import of auto-generated peripheral register definitions.
cfrantz's conversion of the UART to use the auto-generated definitions.

Signed-off-by: Chris Frantz <cfrantz@google.com>